+++ cli/lib/context/context_get_pull_request_url.dart
@@ -1,8 +1,8 @@
 import 'package:stax/context/context.dart';
 import 'package:stax/context/context_git_get_default_remote.dart';
 
-extension ContextGetPrUrl on Context {
-  String? getPrUrl(String targetBranch, String currentBranch) {
+extension ContextGetPullRequestUrl on Context {
+  String? getPullRequestUrl(String baseBranch, String currentBranch) {
     final remote = getPreferredRemote();
     if (remote == null) return null;
     final remoteUrl = git.remoteGetUrl
@@ -14,6 +14,6 @@ extension ContextGetPrUrl on Context {
         .replaceFirstMapped(RegExp(r'git@(.*):'), (m) => 'https://${m[1]}/');
     if (remoteUrl.isEmpty) return null;
 
-    return '${remoteUrl.substring(0, remoteUrl.length - 4)}/compare/$targetBranch...$currentBranch?expand=1';
+    return '${remoteUrl.substring(0, remoteUrl.length - 4)}/compare/$baseBranch...$currentBranch?expand=1';
   }
 }